Fix resolving relative namespaces
Relative namespaces like {{..:myfile.svg}} are currently not resolved by
svgembed, but they should, as this is what happens for other image
types.

It doesn't work because resolve_mediaid() is not called in the xhtml
render path.

To fix it, move resolve_mediaid() to a common place and use it for for
both xhtml and metadata render paths.

Signed-off-by: Tobias Deiminger <tdmg@linutronix.de>